refactoring jsons

+++ src/products/products.controller.ts
@@ -6,27 +6,30 @@ import Product from "../../DTOs/Products/Product";
 
 @Controller('products')
 export class ProductsController {
-    constructor(private readonly productsService: ProductsService) {}
+   constructor(private readonly productsService: ProductsService) {}
 
-    @Get()
-    async getProducts(@Res() res: Response): Promise<Response<Array<Product>>> {
-        return res.status(200).json({ data: await this.productsService.get() }).end();
-    }
+   @Get()
+   async getProducts(@Res() res: Response): Promise<Response<Array<Product>>> {
+      return res.status(200).json({ data: await this.productsService.get() }).end();
+   }
 
-    @Post('register')
-    async productsRegister(@Body() data: PRegisterDTO, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
-        return res.status(201).json({ data: await this.productsService.register(data, token.replace('Bearer ', '')) });
-    }
+   @Post('register')
+   async productsRegister(@Body() data: PRegisterDTO, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
+      if (!token) return res.status(401).end();
+      return res.status(201).json({ data: await this.productsService.register(data, token.replace('Bearer ', '')) });
+   }
 
-    @Put('update/:id')
-    async updateProduct(@Param('id') id: string, @Body() data: PRegisterDTO, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
-        const updated = await this.productsService.update(data, Number(id), token.replace("Bearer ", ''));
-        return res.status(202).json({ data: updated }).end();
-    }
+   @Put('update/:id')
+   async updateProduct(@Param('id') id: string, @Body() data: PRegisterDTO, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
+      if (!token) return res.status(401).end();
+      const updated = await this.productsService.update(data, Number(id), token.replace("Bearer ", ''));
+      return res.status(202).json({ data: updated }).end();
+   }
 
-    @Delete('delete/:id')
-    async deleteProduct(@Param('id') id: string, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
-        return res.status(202).json({ data: await this.productsService.delete(Number(id), token.replace('Bearer ', '')) });
-    }
+   @Delete('delete/:id')
+   async deleteProduct(@Param('id') id: string, @Headers('authorization') token: string, @Res() res: Response): Promise<Response<Product>> {
+      if (!token) return res.status(401).end();
+      return res.status(202).json({ data: await this.productsService.delete(Number(id), token.replace('Bearer ', '')) });
+   }
 }
